refactor to prepare for wasm variant build
