+++ src/vtvalidate-nan.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <nan.h>
+
+namespace vtvalidate_nan {
+NAN_METHOD(isValid);
+}
